+++ r_exec/view.inline.cpp
@@ -147,11 +147,11 @@ inline void View::init(SyncMode sync,
   code_[VIEW_OID].atom_ = GetOID();
   reset_ctrl_values();
 
-  code(VIEW_SYNC) = Atom::Float(sync);
+  code(VIEW_SYNC) = Atom::Float((float32)sync);
   code(VIEW_IJT) = Atom::IPointer(code(VIEW_OPCODE).getAtomCount() + 1);
   r_code::Utils::SetTimestamp<View>(this, VIEW_IJT, ijt);
   code(VIEW_SLN) = Atom::Float(sln);
-  code(VIEW_RES) = res < 0 ? Atom::PlusInfinity() : Atom::Float(res);
+  code(VIEW_RES) = res < 0 ? Atom::PlusInfinity() : Atom::Float((float32)res);
   code(VIEW_HOST) = Atom::RPointer(0);
   code(VIEW_ORG) = origin ? Atom::RPointer(1) : Atom::Nil();
